Adding calendar and changing layout

+++ app/views.py
@@ -1,36 +1,38 @@
 from django.shortcuts import render, redirect, HttpResponse
 from app.forms import SuggestionForm, SignUpForm
-from app.models import Suggestion, TimeSlot, Date, Appointment
+from app.models import Suggestion, TimeSlot, Appointment, Professor, Student, ProfessorStudent
 from django.contrib.auth import login, authenticate
 from django.contrib.auth.models import User
 from django.contrib.auth.forms import UserCreationForm
 import json
 from django.http import JsonResponse
+from app.helpers.email_helper import MassEmailSender 
 
 # Create your views here.
 
 def index(request):
-
-    user = request.user
-
-    if user.is_authenticated:
-        return render(request, 'index.html', {'title' : 'Home', 'message' : 'Hello, and welcome to my site'})
-    else:
-        return redirect("/accounts/login/")
-
-    
-
-
+    data = {'title' : 'Home'}
+    status = request.session['status'];
+    data['status'] = status
+    if (status == 'student'):
+        data['professor_email'] = request.session['professor_email']
+    return render(request, 'index.html', data)
+ 
 def contact(request):
+    return render(request, 'contact.html', {'title' : 'Contact'})
 
-    return render(request, 'contact.html', {'title' : 'Contact', 'message' : 'This page contains my contact info'})
+def send_remainder_emails(request):
 
+    massEmailSender = MassEmailSender();
+    massEmailSender.sendAllEmails();
+
+    return JsonResponse({"result" : "Ok"})
 
 def suggestions(request):
 
     if request.method == 'GET':
         # If the method is get, then just display the suggestion page
-        return render(request, 'suggestions.html', {'title' : 'Suggestions', 'form' : SuggestionForm(), 'message' : 'Suubmit a new Suggestion'})
+        return render(request, 'suggestions.html', {'title' : 'Suggestions', 'form' : SuggestionForm()})
     else:
         # If the method is post, then store the value
         form = SuggestionForm(request.POST);
@@ -55,7 +57,7 @@ def delete_suggestion(request, id):
 
     all = Suggestion.objects.all();
 
-    return render(request, 'all_suggestions.html', {"title" : "Suggestions", "suggestions" : all, 'message' : 'One Suggestion deleted'})
+    return render(request, 'all_suggestions.html', {"title" : "Suggestions", "suggestions" : all})
 
 
 def legal_information(request):
@@ -95,15 +97,23 @@ def profile(request):
 def timeslots(request):
 
     if request.method == 'GET':
+        slots = TimeSlot.objects.order_by("-date", "-start_time")
         # If the method is get, then just display the suggestion page
-        return render(request, 'professor_view/add_time_slot.html', {'title' : 'Add time slot'})
+        return render(request, 'professor_view/add_time_slots_2.html', {'title' : 'Add time slot', "timeslots" : slots})
     else:
         # slots = json.loads(request.POST.get("slots", ""))
         startTimes = request.POST.getlist("startTimes[]")
         endTimes = request.POST.getlist("endTimes[]")
-        professor_name = request.POST.get("professor_name")
         professor_email = request.POST.get("professor_email")
-        date = request.POST.get("date")        
+        prof = None
+        try:
+            prof = Professor.objects.get(email = professor_email);
+        except Professor.DoesNotExist:
+            return JsonResponse({'result' : "Erro"})
+        date = request.POST.get("date") 
+        professor_name = request.user.first_name + " " + request.user.last_name;      
+        if prof != None:
+              professor_name = prof.title + " " + professor_name
         for i in range(len(startTimes)):
             start = startTimes[i]
             end = endTimes[i]
@@ -118,7 +128,7 @@ def all_timeslots(request):
 
     slots = TimeSlot.objects.order_by("-date", "-start_time")
 
-    return render(request, 'common/all_time_slots.html', {'timeslots' : slots, "title" : "All TimeSlots"})
+    return render(request, 'common/all_time_slots_2.html', {'timeslots' : slots, "title" : "All TimeSlots"})
 
 
 def delete_timeslot(request, id):
@@ -141,15 +151,25 @@ def appointments(request):
 
     else:
         # slots = json.loads(request.POST.get("slots", ""))
-        appointment = Appointment()
-        appointment.save_data(request.POST, request.user)
-        appointment.save()
+        appointments = Appointment.objects.filter(user_email = request.user.email);
+        if (appointments.count() > 0):
+            return JsonResponse({"result" : "Error"});
+        appointment = Appointment();
+        profRel = ProfessorStudent.objects.get(student_email = request.user.email);
+        prof = Professor.objects.get(email = profRel.professor_email);
+        profUser = User.objects.get(email = profRel.professor_email)
+        appointment.save_data(request.POST, request.user, profUser, prof);
+        appointment.save();
         TimeSlot.objects.get(id=request.POST['timeslot_id']).delete()
         return JsonResponse({'result' : 'Ok'})
 
     
 def all_appointments(request):
-    a = Appointment.objects.filter(user_email=request.user.email)
+    a = []
+    if (request.session['status'] == "student"):
+        a = Appointment.objects.filter(user_email=request.user.email)
+    if (request.session['status'] == "professor"):
+        a = Appointment.objects.filter(professor_email = request.user.email)
     return render(request, 'common/all_appointments.html', {'appointments' : a, "title" : "All Appointments"})
